Refactor code to move state into Reservations from DatePicker
Allows for better code readability and simplicity. Initially,
kept state within multiple components in an attempt to reduce
the number of re-renders the service would need to perform. After
further thought, decided that the number of re-renders was less
important than being able to write simple and readable code, thus this
and future refactors.